requirements.txt
asgiref==3.7.2
Django==4.2.9
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
